pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Agentic search demo using the WANDS dataset"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.0",
    "searcharray>=0.0.73",
    "pystemmer>=3.0.0",
    "pyarrow>=14.0.0",
    "numpy>=1.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/search_agent"]
